requirements.txt
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.32.0
coverage==7.0.5
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.5
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
jwcrypto==1.4.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
wrapt==1.14.1
